+++ lib/src/runner.dart
@@ -14,6 +14,7 @@ import 'package:term_glyph/term_glyph.dart' as glyph;
 
 import 'io.dart';
 import 'migrators/division.dart';
+import 'migrators/media_logic.dart';
 import 'migrators/module.dart';
 import 'migrators/namespace.dart';
 import 'migrators/strict_unary.dart';
@@ -54,6 +55,7 @@ class MigratorRunner extends CommandRunner<Map<Uri, String>> {
       ..addFlag('version',
           help: 'Print the version of the Sass migrator.', negatable: false);
     addCommand(DivisionMigrator());
+    addCommand(MediaLogicMigrator());
     addCommand(ModuleMigrator());
     addCommand(NamespaceMigrator());
     addCommand(StrictUnaryMigrator());