+++ test/migrators/migrator_dart_test.dart
@@ -1,4 +1,4 @@
-// Copyright 2021 Google LLC
+// Copyright 2022 Google LLC
 //
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file or at
@@ -7,5 +7,9 @@
 import '../utils.dart';
 
 main() {
+  testMigrator("division");
+  testMigrator("media_logic");
+  testMigrator("module");
   testMigrator("namespace");
+  testMigrator("strict_unary");
 }